WIP: Working on a new idea for the authorization process
This commit is more for backup purposes. Something funky
is happening with IDEA and/or gradle.

The build fails like a complete idiot. When I want it to
blast me with exception and crash info it is not
happening, but when I try to have a nice life it bugs me
like a mad man.